% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkFolder}
\alias{checkFolder}
\title{checkFolder}
\usage{
checkFolder(n = 100)
}
\arguments{
\item{n}{n}
}
\value{
value
}
\description{
Check whether folder name from 1 to n exists
}
\details{
checkFolder
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

