% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkMaxFolder}
\alias{checkMaxFolder}
\title{checkMaxFolder}
\usage{
checkMaxFolder(n = 20)
}
\arguments{
\item{n}{max integer for folder name}
}
\value{
value
}
\description{
What is the largest value of folder name. (Folder name must be integer)
}
\details{
checkMaxFolder
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

