% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkNewStart}
\alias{checkNewStart}
\title{checkNewStart}
\usage{
checkNewStart(command_args, create_dir = TRUE)
}
\arguments{
\item{command_args}{folder name}

\item{create_dir}{logical. If the folder doesn't exist, create a new directory with name \code{command_args}}
}
\value{
value
}
\description{
Check whether there is folder named as \code{command_args}
}
\details{
checkNewStart
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

