% Generated by roxygen2 (4.0.1): do not edit by hand
\name{commandArgsParser}
\alias{commandArgsParser}
\title{commandArgsParser}
\usage{
commandArgsParser()
}
\value{
value
}
\description{
commandArgsParser
}
\details{
This function will check whether the commandArgs contain "-ID". If it does, the argument will be return, if not, it will return the max folder number +1. that's mean if there is not argument, it will use a new folder used in beginning of runme.r to determine where to store the mcmc reuslt
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

