% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compareMCMC}
\alias{compareMCMC}
\title{compareMCMC}
\usage{
compareMCMC(x, tail, plot = FALSE, pdf = FALSE)
}
\arguments{
\item{x}{x}

\item{tail}{tail}

\item{plot}{Logical. Whether to have plot}

\item{pdf}{Logical}
}
\value{
value
}
\description{
compareMCMC
}
\details{
compareMCMC
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

