% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeFittedValue}
\alias{computeFittedValue}
\title{computeFittedValue}
\usage{
computeFittedValue(x, data, ...)
}
\arguments{
\item{x}{x}

\item{data}{data}

\item{...}{...}
}
\value{
value
}
\description{
computeFittedValue
}
\details{
computeFittedValue
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

