% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeFittedValue.simmen}
\alias{computeFittedValue.simmen}
\title{computeFittedValue.simmen}
\usage{
\method{computeFittedValue}{simmen}(x, data, ...)
}
\arguments{
\item{x}{x}

\item{data}{data}

\item{...}{...}
}
\value{
value
}
\description{
Compute the fitted value of the model.
}
\details{
computeFittedValue.simmen
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

