% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeFV}
\alias{computeFV}
\title{computeFV}
\usage{
computeFV(phi, lambda, data)
}
\arguments{
\item{phi}{phi}

\item{lambda}{lambda}

\item{data}{data}
}
\value{
value
}
\description{
Compute the fitted value of the model.
}
\details{
computeFV
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

