\name{computeNetworkSummary}
\alias{computeNetworkSummary}
\title{computeNetworkSummary}
\usage{
computeNetworkSummary(seq_m, D)
}
\arguments{
  \item{seq_m}{meeting sequence}

  \item{D}{adjacency matrix}
}
\value{
value
}
\description{
computeNetworkSummary
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

