% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeNetworkSummary_cxx}
\alias{computeNetworkSummary_cxx}
\title{computeNetworkSummary_cxx}
\usage{
computeNetworkSummary_cxx(seq_m, D)
}
\arguments{
\item{seq_m}{seq_m}

\item{D}{D}
}
\value{
value
}
\description{
computeNetworkSummary_cxx
}
\details{
computeNetworkSummary_cxx
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

