% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeSummaryTable}
\alias{computeSummaryTable}
\title{computeSummaryTable}
\usage{
computeSummaryTable(x, tail)
}
\arguments{
\item{x}{x}

\item{tail}{tail}
}
\value{
value
}
\description{
computeSummaryTable
}
\details{
computeSummaryTable
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

