% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeSummaryTableFromDrive}
\alias{computeSummaryTableFromDrive}
\title{computeSummaryTableFromDrive}
\usage{
computeSummaryTableFromDrive(path, tail, ...)
}
\arguments{
\item{path}{path}

\item{tail}{tail}

\item{...}{...}
}
\value{
value
}
\description{
computeSummaryTableFromDrive
}
\details{
computeSummaryTableFromDrive
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

