% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DrawSeqSample}
\alias{DrawSeqSample}
\title{DrawSeqSample}
\usage{
DrawSeqSample(x, p = 0.01)
}
\arguments{
\item{x}{x}

\item{p}{p}
}
\value{
value
}
\description{
Draw random sample of meeting sequence
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

