\name{extractTail}
\alias{extractTail}
\title{extractTail}
\usage{
extractTail(x, tail = 0)
}
\arguments{
  \item{x}{x}

  \item{tail}{default = 0}
}
\value{
value
}
\description{
Flexible version of tail
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
link \code{\link{tail}}
}

