% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findLikelihood}
\alias{findLikelihood}
\title{findLikelihood}
\usage{
findLikelihood(data, last_out)
}
\arguments{
\item{data}{data}

\item{last_out}{last_out}
}
\value{
value
}
\description{
Compute the likelihood basic on the parameter of last iteration.
}
\details{
findLikelihood
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

