% Generated by roxygen2 (4.0.1): do not edit by hand
\name{findplotIndex}
\alias{findplotIndex}
\title{findplotIndex}
\usage{
findplotIndex(x)
}
\arguments{
\item{x}{x}
}
\value{
value
}
\description{
findplotIndex
}
\details{
findplotIndex
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

