% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genFulleByNetwork}
\alias{genFulleByNetwork}
\title{genFulleByNetwork}
\usage{
genFulleByNetwork(data, e)
}
\arguments{
\item{data}{data}

\item{e}{e}
}
\value{
value
}
\description{
Return e_i e_j and |e_i - e_j|
}
\details{
genFulleByNetwork
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

