% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genFullGroupIndex}
\alias{genFullGroupIndex}
\title{genFullGroupIndex}
\usage{
genFullGroupIndex(data)
}
\arguments{
\item{data}{data}
}
\value{
value
}
\description{
genFullGroupIndex
}
\details{
genFullGroupIndex
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

