% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genFullPositionMatrix}
\alias{genFullPositionMatrix}
\title{genFullPositionMatrix}
\usage{
genFullPositionMatrix(data)
}
\arguments{
\item{data}{data}
}
\value{
value
}
\description{
genFullPositionMatrix
}
\details{
genFullPositionMatrix
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

