\name{genLogDetGamma}
\alias{genLogDetGamma}
\title{genLogDetGamma}
\usage{
genLogDetGamma(W_list, lambda)
}
\arguments{
  \item{W_list}{a list of the weighting matrix}

  \item{lambda}{a vector of lambda}
}
\value{
value
}
\description{
Compute the log of the determinant of I - \\sum_i
\\lambda_i W_i
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

