% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genPoly}
\alias{genPoly}
\title{genPoly}
\usage{
genPoly(e_i, e_j, delta)
}
\arguments{
\item{e_i}{e_i}

\item{e_j}{e_j}

\item{delta}{delta}
}
\value{
value
}
\description{
genPoly
}
\details{
genPoly
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

