% Generated by roxygen2 (4.0.1): do not edit by hand
\name{genUniqueID}
\alias{genUniqueID}
\title{genUniqueID}
\usage{
genUniqueID()
}
\value{
value
}
\description{
gen an unique ID
}
\details{
it depends on the time and a 8 digits random integer
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

