% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getParameterMatrix}
\alias{getParameterMatrix}
\title{getParameterMatrix}
\usage{
getParameterMatrix(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}
}
\value{
value
}
\description{
getParameterMatrix
}
\details{
getParameterMatrix
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

