\name{getParameterMatrix.simmen}
\alias{getParameterMatrix.simmen}
\title{getParameterMatrix.simmen}
\usage{
\method{getParameterMatrix}{simmen}(x, tail, ...)
}
\arguments{
  \item{x}{x}

  \item{tail}{tail}

  \item{...}{...}
}
\value{
value
}
\description{
Get a matrix of all parameters
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

