\name{loadLastObject}
\alias{loadFirstObject}
\alias{loadLastObject}
\alias{readAllLastObject}
\title{loadLastObject}
\usage{
loadLastObject(foldername)
}
\arguments{
  \item{foldername}{foldername}
}
\value{
value
}
\description{
load first/last object
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

