\name{loglikelihood_outcome_vector_multi}
\alias{loglikelihood_outcome_vector_multi}
\title{loglikelihood_outcome_vector_multi}
\usage{
loglikelihood_outcome_vector_multi(e_matrix, new_e, lambda, phi, Y, X, WY,
  network_id, rho)
}
\arguments{
  \item{e_matrix}{e_matrix}

  \item{new_e}{new_e}

  \item{lambda}{lambda}

  \item{phi}{phi}

  \item{Y}{Y}

  \item{X}{X}

  \item{WY}{WY}

  \item{network_id}{network_id}

  \item{rho}{rho}
}
\value{
value
}
\description{
loglikelihood_outcome_vector_multi
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

