\name{loglikelihood_SNF}
\alias{loglikelihood_SNF}
\title{loglikelihood_SNF}
\usage{
loglikelihood_SNF(y, x1, x2, delta, y_not)
}
\arguments{
  \item{y}{indicator of whether i and j is connected}

  \item{x1}{variables of i}

  \item{x2}{variables of j}

  \item{delta}{parameters}

  \item{y_not}{complement of y}
}
\value{
value of log likelihood
}
\description{
loglikelihood_SNF
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

