% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logLikeliMultiNormalSigma}
\alias{logLikeliMultiNormalSigma}
\title{logLikeliMultiNormalSigma}
\usage{
logLikeliMultiNormalSigma(x, sigma)
}
\arguments{
\item{x}{x}

\item{sigma}{sigma}
}
\value{
value
}
\description{
logLikeliMultiNormalSigma
}
\details{
logLikeliMultiNormalSigma
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

