\name{merge.network_formation.mcmc}
\alias{merge.network_formation.mcmc}
\title{merge.network_formation.mcmc}
\usage{
\method{merge}{network_formation.mcmc}(x, y, ...)
}
\arguments{
  \item{x}{First object to merge with}

  \item{y}{Second object to merge with}

  \item{...}{not used}
}
\value{
A new network_formation.mcmc object
}
\description{
Description merge.network_formation.mcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

