% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.simmen}
\alias{merge.simmen}
\title{merge.simmen}
\usage{
\method{merge}{simmen}(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{...}
}
\value{
value
}
\description{
merge.simmen
}
\details{
merge.simmen
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

