\name{merge.SNF}
\alias{merge.SNF}
\title{merge.SNF}
\usage{
\method{merge}{SNF}(x, y, ...)
}
\arguments{
  \item{x}{First object to merge with}

  \item{y}{Second object to merge with}

  \item{...}{not used}
}
\value{
A new SNF object
}
\description{
Description merge.SNF
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

