\name{merge.SNF.static.mcmc}
\alias{merge.SNF.static.mcmc}
\title{merge.SNF.static.mcmc}
\usage{
\method{merge}{SNF.static.mcmc}(x, y, ...)
}
\arguments{
  \item{x}{First object to merge with}

  \item{y}{Second object to merge with}

  \item{...}{not used}
}
\value{
A new SNF.static.mcmc object
}
\description{
merge.SNF.static.mcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

