\name{mergeTable}
\alias{mergeTable}
\title{mergeTable}
\usage{
mergeTable(x, y, seperator = c("friends_", "studymates_"),
  title = c("Friends", "Studymates"))
}
\arguments{
  \item{x}{x}

  \item{y}{y}

  \item{seperator}{seperator}

  \item{title}{title}
}
\value{
value
}
\description{
mergeTable
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

