\name{my.rtnorm}
\alias{my.rtnorm}
\title{my.rtnorm}
\usage{
my.rtnorm(mean = 0, sd = 1, a, b)
}
\arguments{
  \item{mean}{mean}

  \item{sd}{sd}

  \item{a}{a}

  \item{b}{b}
}
\value{
value
}
\description{
Generate random variable from truncated normal
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

