\name{network_formation.maxLik}
\alias{network_formation.maxLik}
\title{network_formation.maxLik}
\usage{
network_formation.maxLik(data, ...)
}
\arguments{
  \item{data}{data}

  \item{...}{not used}
}
\value{
network_formation object
}
\description{
Description network_formation.maxLik
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

