\name{network_formation.mcmc}
\alias{network_formation.mcmc}
\title{network_formation.mcmc}
\usage{
network_formation.mcmc(data, m = 1000, last_estimation, ...)
}
\arguments{
  \item{data}{data}

  \item{m}{number of iteration}

  \item{last_estimation}{previous estimation object}

  \item{...}{not used}
}
\value{
network_formation object
}
\description{
Description network_formation.mcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

