\name{network_formation}
\alias{network_formation}
\title{network_formation}
\usage{
network_formation(data, method = c("maxLik", "mcmc"), ...)
}
\arguments{
  \item{data}{data}

  \item{method}{Estimation method, either "maxLik" or
  "mcmc". Default is "maxLik"}

  \item{...}{others argument. See
  \link{network_formation.mcmc}}
}
\value{
network_formation object
}
\description{
Description network_formation
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

