% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotmcmc.default}
\alias{plotmcmc.default}
\title{plotmcmc.default}
\usage{
\method{plotmcmc}{default}(x, tail = 0, ...)
}
\arguments{
\item{x}{x}

\item{tail}{tail}

\item{...}{...}
}
\value{
value
}
\description{
plotmcmc.default
}
\details{
plotmcmc.default
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

