% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotmcmc}
\alias{plotmcmc}
\alias{plotmcmc2}
\title{plotmcmc}
\usage{
plotmcmc(x, ...)

plotmcmc2(x, ...)
}
\arguments{
\item{x}{x}

\item{...}{...}
}
\value{
value
}
\description{
plotmcmc
}
\details{
plotmcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

