\name{plotmcmc.simmen}
\alias{plotmcmc.simmen}
\title{plotmcmc.simmen}
\usage{
\method{plotmcmc}{simmen}(x, tail = -0.2, ...)
}
\arguments{
  \item{x}{x}

  \item{tail}{tail}

  \item{...}{...}
}
\value{
value
}
\description{
plotmcmc.simmen
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

