% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotmcmc2.default}
\alias{plotmcmc2.default}
\title{plotmcmc2.default}
\usage{
\method{plotmcmc2}{default}(x, name, file = "mcmc_plot.pdf", tail, ...)
}
\arguments{
\item{x}{x}

\item{name}{name}

\item{file}{file}

\item{tail}{tail}

\item{...}{...}
}
\value{
value
}
\description{
plotmcmc2.default
}
\details{
plotmcmc2.default
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

