\name{plotmcmc_byid}
\alias{plotmcmc_byid}
\title{plotmcmc_byid}
\usage{
plotmcmc_byid(x, id, mean, min, max)
}
\arguments{
  \item{x}{x}

  \item{id}{id}

  \item{mean}{mean}

  \item{min}{min}

  \item{max}{max}
}
\value{
value
}
\description{
plotmcmc_byid
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

