% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotmcmc_ma}
\alias{plotmcmc_ma}
\title{plotmcmc_ma}
\usage{
plotmcmc_ma(x, tail, ...)
}
\arguments{
\item{x}{x}

\item{tail}{tail}

\item{...}{...}
}
\value{
value
}
\description{
plotmcmc_ma
}
\details{
plotmcmc_ma
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

