\name{priorFunction}
\alias{priorFunction}
\title{priorFunction}
\usage{
priorFunction(x, type = c("normal", "invgamma", "uniform", "trunc_normal"))
}
\arguments{
  \item{x}{x}

  \item{type}{"normal","invgamma","uniform","trunc_normal"}
}
\value{
value
}
\description{
priorFunction
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

