% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readAllObject}
\alias{readAllObject}
\title{readAllObject}
\usage{
readAllObject(path, except)
}
\arguments{
\item{path}{path}

\item{except}{vector of filenames for exclusion}
}
\value{
value
}
\description{
readAllObject
}
\details{
readAllObject
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

