\name{samplingFunction}
\alias{samplingFunction}
\title{samplingFunction}
\usage{
samplingFunction(beta, tau, type = c("normal", "invgamma", "uniform",
  "trunc_normal"))
}
\arguments{
  \item{beta}{beta}

  \item{tau}{tau}

  \item{type}{"normal","invgamma","uniform","trunc_normal"}
}
\value{
value
}
\description{
samplingFunction
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

