% Generated by roxygen2 (4.0.1): do not edit by hand
\name{save.mcmc}
\alias{save.mcmc}
\title{save.mcmc}
\usage{
save.mcmc(x, i, foldername)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{foldername}{foldername}
}
\value{
value
}
\description{
save.mcmc
}
\details{
save.mcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

