\name{SNF}
\alias{SNF}
\title{SNF}
\usage{
SNF(data, method = c("static.maxLik", "static.mcmc", "dynamic.mcmc"), ...)
}
\arguments{
  \item{data}{data}

  \item{method}{Estimation method, either
  "static.maxLik","static.mcmc","dynamic.mcmc". Default is
  "static.maxLik"}

  \item{...}{others argument. See \link{SNF.static.maxLik},
  \link{SNF.static.mcmc}, \link{SNF.dynamic.mcmc}}
}
\value{
SNF object
}
\description{
Description Strategy Network Formation
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

