\name{SNF.static.maxLik}
\alias{SNF.static.maxLik}
\title{SNF.static.maxLik}
\usage{
SNF.static.maxLik(data, ...)
}
\arguments{
  \item{data}{data}

  \item{...}{not used}
}
\value{
SNF.static.maxLik object
}
\description{
Description SNF.static.maxLik
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

