\name{SNF.static.mcmc}
\alias{SNF.static.mcmc}
\title{SNF.static.mcmc}
\usage{
SNF.static.mcmc(data, m = 1000, last_estimation, ...)
}
\arguments{
  \item{data}{data}

  \item{m}{number of iteration}

  \item{last_estimation}{previous estimation object}

  \item{...}{not used}
}
\description{
Description SNF.static.mcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

