\name{SplitTable}
\alias{SplitTable}
\title{SplitTable}
\usage{
SplitTable(x, seperator = c("friends_", "studymates_"), title = c("Friends",
  "Studymates"))
}
\arguments{
  \item{x}{x}

  \item{seperator}{seperator}

  \item{title}{title}
}
\value{
value
}
\description{
SplitTable
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

