\name{summary.network_formation.maxLik}
\alias{summary.network_formation.maxLik}
\title{summary.network_formation.maxLik}
\usage{
\method{summary}{network_formation.maxLik}(object, ...)
}
\arguments{
  \item{object}{network_formation object}

  \item{...}{not used}
}
\value{
A summary table
}
\description{
Description Create a summary table
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

