\name{summary.SNF.static.maxLik}
\alias{summary.SNF.static.maxLik}
\title{summary.SNF.static.maxLik}
\usage{
\method{summary}{SNF.static.maxLik}(object, ...)
}
\arguments{
  \item{object}{SNF.static.maxLik object}

  \item{...}{not used}
}
\value{
A summary table
}
\description{
Description Create a summary table
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

