% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transform_e2}
\alias{transform_e2}
\title{transform_e2}
\usage{
transform_e2(data, e)
}
\arguments{
\item{data}{data}

\item{e}{e}
}
\value{
value
}
\description{
Apply transform_e2_by_each to each network
}
\details{
transform_e2
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

