\name{uniqueRow}
\alias{uniqueRow}
\title{uniqueRow}
\usage{
uniqueRow(x, number_of_col = "all", percentage = TRUE)
}
\arguments{
  \item{x}{x}

  \item{number_of_col}{number_of_col}

  \item{percentage}{percentage}
}
\value{
value
}
\description{
uniqueRow
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

