\name{update_e_internal_single}
\alias{update_e_internal_single}
\title{update_e_internal_single}
\usage{
update_e_internal_single(location_index_all, location_index1, location_index2,
  e, new_e, xb1, xb2, delta, prob_old1, prob_old2, ystar1, ystar2, e_i, e_j,
  uniform_rv, prob_diff, sigma2)
}
\arguments{
  \item{location_index_all}{location_index_all}

  \item{location_index1}{location_index1}

  \item{location_index2}{location_index2}

  \item{e}{e}

  \item{new_e}{new_e}

  \item{xb1}{xb1}

  \item{xb2}{xb2}

  \item{delta}{delta}

  \item{prob_old1}{prob_old1}

  \item{prob_old2}{prob_old2}

  \item{ystar1}{ystar1}

  \item{ystar2}{ystar2}

  \item{e_i}{e_i}

  \item{e_j}{e_j}

  \item{uniform_rv}{uniform_rv}

  \item{prob_diff}{prob_diff}

  \item{sigma2}{sigma2}
}
\value{
value
}
\description{
update_e_internal_single
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

