\name{computeSummaryTableFromDrive}
\alias{computeSummaryTableFromDrive}
\title{computeSummaryTableFromDrive}
\usage{
computeSummaryTableFromDrive(path)
}
\arguments{
  \item{path}{path}
}
\value{
value
}
\description{
computeSummaryTableFromDrive
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

