\name{drawingPhi}
\alias{drawingPhi}
\title{drawingPhi}
\usage{
drawingPhi(lambda, rho_e, data_matrix)
}
\arguments{
  \item{lambda}{lambda}

  \item{rho_e}{rho_e}

  \item{data_matrix}{data_matrix}
}
\value{
value
}
\description{
drawingPhi
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

