% Generated by roxygen2 (4.0.1): do not edit by hand
\name{extractTail}
\alias{extractTail}
\title{extractTail}
\usage{
extractTail(x, tail = 0)
}
\arguments{
\item{x}{x}

\item{tail}{default = 0}
}
\value{
value
}
\description{
Flexible version of tail
}
\details{
if  |x| < 1, x will be considered as a percentage, otherwise, it will be return tail(x,tail)
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}
\seealso{
link \code{\link{tail}}
}

