\name{find_normal_conditional_dist}
\alias{find_normal_conditional_dist}
\title{find_normal_conditional_dist}
\usage{
find_normal_conditional_dist(a, i, j, Sigma)
}
\arguments{
  \item{a}{mean}

  \item{i}{index or location of x1}

  \item{j}{index or location of x2}

  \item{Sigma}{Covariance matrix}
}
\value{
value
}
\description{
Compute the conditional normal distribution
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

