% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getParameterMatrix.simmen}
\alias{getParameterMatrix.simmen}
\title{getParameterMatrix.simmen}
\usage{
\method{getParameterMatrix}{simmen}(x, tail, ...)
}
\arguments{
\item{x}{x}

\item{tail}{tail}

\item{...}{can specify the name of one parameter(e.g. getParameterMatrix(x,tail=50,a="phi")). All parameter if missing.}
}
\value{
value
}
\description{
Get a matrix of all parameters
}
\details{
getParameterMatrix.simmen
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

