\name{dinvgamma}
\alias{dinvgamma}
\alias{rinvgamma}
\title{invgamma}
\usage{
dinvgamma(x, shape, scale = 1)

rinvgamma(n, shape, scale = 1)
}
\arguments{
  \item{x}{x}

  \item{n}{n}

  \item{shape}{shape}

  \item{scale}{scale}
}
\value{
value
}
\description{
Inverse gamma Distribution
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

