\name{lik_grad_single_network_formation}
\alias{lik_grad_single_network_formation}
\title{lik_grad_single_network_formation}
\usage{
lik_grad_single_network_formation(y, x1, x2, delta, y_not)
}
\arguments{
  \item{y}{indicator of whether i and j is connected}

  \item{x1}{variables of i}

  \item{x2}{variables of j}

  \item{delta}{parameters}

  \item{y_not}{complement of y}
}
\value{
value of gradient of log likelihood
}
\description{
Description lik_grad_single_network_formation
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

