% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadLastObject}
\alias{loadFirstObject}
\alias{loadLastObject}
\alias{readAllLastObject}
\title{loadLastObject}
\usage{
loadLastObject(foldername)

loadFirstObject(foldername)

readAllLastObject()
}
\arguments{
\item{foldername}{foldername}
}
\value{
value
}
\description{
load first/last object
}
\details{
load first/last/all last object. \code{readAllLastObject} will read all last object from every folder (with name from 1 to 20)
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

