% Generated by roxygen2 (4.0.1): do not edit by hand
\name{loadMergeMCMC}
\alias{loadMergeMCMC}
\title{loadMergeMCMC}
\usage{
loadMergeMCMC(max = 40, folder_exist)
}
\arguments{
\item{max}{max number of merge}

\item{folder_exist}{vector of names of folder}
}
\value{
value
}
\description{
loadMergeMCMC
}
\details{
loadMergeMCMC
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

