\name{logLikeliMultiNormalSigma}
\alias{logLikeliMultiNormalSigma}
\title{logLikeliMultiNormalSigma}
\usage{
logLikeliMultiNormalSigma(x, sigma)
}
\arguments{
  \item{x}{x}

  \item{sigma}{sigma}
}
\value{
value
}
\description{
logLikeliMultiNormalSigma
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

