% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.SNF.dynamic.mcmc}
\alias{merge.SNF.dynamic.mcmc}
\title{merge.SNF.dynamic.mcmc}
\usage{
\method{merge}{SNF.dynamic.mcmc}(x, y, ...)
}
\arguments{
\item{x}{First object to merge with}

\item{y}{Second object to merge with}

\item{...}{not used}
}
\value{
A new SNF.dynamic.mcmc object
}
\description{
merge.SNF.dynamic.mcmc
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

