\name{plotmcmc.default}
\alias{plotmcmc.default}
\title{plotmcmc.default}
\usage{
\method{plotmcmc}{default}(x, tail = 0, ...)
}
\arguments{
  \item{x}{x}

  \item{tail}{tail}

  \item{...}{...}
}
\value{
value
}
\description{
plotmcmc.default
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

