\name{removeFileSizeZero}
\alias{removeFileSizeZero}
\title{removeFileSizeZero}
\usage{
removeFileSizeZero(folder_name)
}
\arguments{
  \item{folder_name}{folder_name}
}
\value{
value
}
\description{
Remover file with size zero(corrupted files)
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

