\name{SNF}
\alias{SNF}
\title{Strategy Network Formation}
\usage{
SNF(m, data, last_estimation, update_tau = TRUE, tau = 0.005)
}
\arguments{
  \item{m}{m}

  \item{data}{data}

  \item{last_estimation}{last_estimation}

  \item{update_tau}{update_tau}

  \item{tau}{tau}
}
\value{
SNF object
}
\description{
Description Strategy Network Formation
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

