% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SNF}
\alias{SNF}
\alias{SNF.dynamic.mcmc}
\alias{SNF.static.maxLik}
\alias{SNF.static.mcmc}
\title{SNF}
\usage{
SNF(data, method = c("static.maxLik", "static.mcmc", "dynamic.mcmc"), ...)

SNF.static.maxLik(data, ...)

SNF.static.mcmc(data, m = 1000, last_estimation, ...)

SNF.dynamic.mcmc(m, data, last_estimation, update_tau = TRUE, tau = 0.005)
}
\arguments{
\item{data}{data}

\item{method}{Estimation method, either "static.maxLik","static.mcmc","dynamic.mcmc". Default is "static.maxLik"}

\item{m}{m}

\item{last_estimation}{last_estimation}

\item{update_tau}{update_tau}

\item{tau}{tau}

\item{...}{others argument.}
}
\value{
SNF object
}
\description{
Strategy Network Formation
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

