% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SplitTable}
\alias{SplitTable}
\title{SplitTable}
\usage{
SplitTable(x, seperator = c("friends_", "studymates_"), title = c("Friends",
  "Studymates"))
}
\arguments{
\item{x}{x}

\item{seperator}{seperator}

\item{title}{title}
}
\value{
value
}
\description{
SplitTable
}
\details{
SplitTable
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

