\name{transform_e_by_each}
\alias{transform_e_by_each}
\title{transform_e_by_each}
\usage{
transform_e_by_each(data, e)
}
\arguments{
  \item{data}{data}

  \item{e}{e}
}
\value{
value
}
\description{
transform a vector of error into absolute difference form
}
\author{
TszKin Julian Chan \email{ctszkin@gmail.com}
}

